HopeSolutions/requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.5
psycopg2-binary==2.9.7
sqlparse==0.4.4
typing_extensions==4.8.0
faker==5.0.*
python-decouple==3.8
djangorestframework==3.14.0

